% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.baseline}
\alias{sim.baseline}
\title{Create a baseline signal}
\usage{
sim.baseline(N, mu, sigma, transform = NULL, transform_par = 2,
  distort = NULL, distort_par = 100, output = c(1, 2, 3),
  flow_rate_baseline = NULL, verbose = FALSE)
}
\arguments{
\item{N}{total number of time points over which events are distributed.}

\item{mu}{mean of distribution from which values are sampled (independently per each time point).}

\item{sigma}{standard deviation (see \code{mu}).}

\item{transform}{type of transformation applied to values per each time tick. Can be \code{log}, \code{asinh}. Defaults to \code{NULL}.}

\item{transform_par}{parameter for non-null \code{transform}. For \code{log}, this is the base. For \code{asinh}, this is \code{b} in \code{y := asinh(y/b)} where \code{y} is the vector of original values.}

\item{distort}{time-dependent distortion of the baseline values. Can be \code{sin}, \code{slope}. Defaults to \code{NULL}.}

\item{distort_par}{parameter for non-null \code{distort}. For \code{sin}, this is a vector of (1) amplitude, (2) period. For \code{slope}, this is \code{a} in \code{offset := a*x + mu}, where \code{x} is a vector of time points and \code{offset} is added to the original values.}

\item{output}{specifies type of output. If 1, the function produces a list containing a \code{df} slot with data frame of \code{time} and \code{values} and a \code{params} slot containing parameters of the signal: \code{mu}, \code{sigma}, \code{D} (equal to \code{"gauss"}, since baseline
values are sampled from a Gaussian distribution), \code{transform}, \code{transform_par}, \code{distort}, \code{distort_par} and \code{rate_adjusted} (Boolean, indicated if parameter \code{flow_rate_baseline} was used). If 2, only the data frame (\code{df}) is returned. If 3, only the
\code{value} column (without \code{time}) of the data frame is returned as numeric vector.}

\item{flow_rate_baseline}{for simulating fluorescence signal congruent with flow rate signal, as generated by \code{sim.baseline} (with possible perturbations). When \code{flow_rate_signal} is non-null, this overrides \code{N} and events are distributed in the way specified by the flow rate.}

\item{verbose}{Boolean, indicates whether the total number of events (\code{N}) should be printed when \code{flow_rate_baseline} is non-null.}
}
\value{
See parameter \code{output}
}
\description{
Create a baseline flow rate or fluorescence signal.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
