% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.concatenate}
\alias{sim.concatenate}
\title{Signal concatenation}
\usage{
sim.concatenate(..., reindex = FALSE, contract_time = TRUE,
  flow_rate = NULL, output = c(1, 2, 3))
}
\arguments{
\item{...}{baselines, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{reindex}{Boolean, indicates whether \code{time} values should be re-indexed to 1..N after contraction. Do not use this if the baseline signals have been generated based on a specific flow rate.}

\item{contract_time}{Boolean, indicates whether to contract time values. Since length of measurement gets multiplied by number of signals due to concatenation, contracting time will downsample to preserve the
original length of measurement.
@param output see parameter \code{output} of function \code{sim.baseline}.}
}
\value{
Modified \code{baseline} input.
}
\description{
Concatenates multiple fluorescence signals (for instance, negative and positive population signals) with the same time scale.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
