% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{sim.make_flowFrame}
\alias{sim.make_flowFrame}
\title{Make flowFrame from synthetic signal data}
\usage{
sim.make_flowFrame(..., channel_names, marker_names = NULL,
  guid = "qc_toy")
}
\arguments{
\item{...}{signals, as generated by function \code{sim.baseline}, possibly perturbed by other functions from \code{qctoy}.}

\item{channel_names}{vector of channel names, of length equal to number of signal data objects.}

\item{marker_names}{vector of marker names, of length equal to number of signal data objects. If \code{NULL}, set to \code{channel_names}.}

\item{guid}{GUID descriptor of \code{flowFrame}.}
}
\value{
\code{flowFrame} object.
}
\description{
Creates a flowFrame with given signal in named channels.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
