% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual.R
\name{sim.plot}
\alias{sim.plot}
\title{Plot signal data}
\usage{
sim.plot(data, bin_width = NULL, plot_avg_and_dev = FALSE,
  avg = c("mean", "median"), point_size = 1, y.lim = NULL,
  title = "Flow rate", y.lab = "flow rate", x.lab = "time")
}
\arguments{
\item{data}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{bin_width}{number of time points per bin. If \code{NULL}, no binning.}

\item{avg}{measure of average. Can be \code{"mean"} (then, measure of deviation is SD) or \code{"median"} (then, measure of deviation is MAD).}

\item{point_size}{size of points.}

\item{y.lim}{y-axis (values) limits (vector of length 2).}

\item{title}{plot title.}

\item{y.lab}{y-axis (values) label.}

\item{x.lab}{x-axis (time) label.}
}
\value{
Modified \code{baseline} input.
}
\description{
Plots signal data. The data can be binned into segments spanning equal time parts.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
