% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.prune}
\alias{sim.prune}
\title{Signal manipulation: prune}
\usage{
sim.prune(baseline, i = 1, j = NULL, N = 2, reindex = FALSE)
}
\arguments{
\item{baseline}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{i}{see description above. Defaults to 1.}

\item{j}{see description above. Defaults to \code{NULL}.}

\item{N}{see description above. Defaults to \code{N}.}

\item{reindex}{Boolean, indicates whether \code{time} values should be re-indexed to 1..N after contraction.}
}
\value{
Modified \code{baseline} input.
}
\description{
Prunes signal data. Given a baseline signal, this function separates signal values into segments of size \code{N}, with respect to time. Then, from each segment, \code{i} values (1..N) are removed at random
(unless \code{j} is not \code{NULL}, in which case always remove the \code{j}-th values in the given segment and disregard \code{i}). Crucially, the removed values are only set to \code{NA}, time scale stays in tact.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
