% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{sim.sample.linear_increase_in_signal}
\alias{sim.sample.linear_increase_in_signal}
\title{Synthetic QC data sample: linear increase in fluorescence signal over time}
\usage{
sim.sample.linear_increase_in_signal(out.flow_rate = NULL,
  out.fluorescence = NULL)
}
\arguments{
\item{out.flow_rate}{if not \code{NULL}, the function will asign an object with flow rate signal data to this.}

\item{out.fluorescence}{if not \code{NULL}, the function will asign an object with fluorescence signal data to this.}
}
\value{
\code{flowFrame} object.
}
\description{
Creates a \code{flowFrame} with artificially generated fluorescence channels, for QC evaluation and benchmarking.
}
\seealso{
See other functions from the \code{qctoy} package. Start with \code{sim.baseline}. For plotting, use \code{sim.plot}.
}
