% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.sharp_push}
\alias{sim.sharp_push}
\title{Signal perturbation: sharp push}
\usage{
sim.sharp_push(baseline, range = NULL, range_by_time = TRUE, strength)
}
\arguments{
\item{baseline}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{range}{indices of affected values. Either a start and end index or the whole range. If length of \code{range} is 2 and second value is infinity, the second value is converted to highest possible value. If \code{range} is \code{NULL}, maximum range (1 to max) is taken.}

\item{range_by_time}{Boolean, indicates whether values in \code{range} correspond to value of \code{time} of signal, rather than row indices. This is recommended, since time values for fluorescence signals with non-uniform flow rate are also not uniform.}

\item{strength}{linear term in offsetting selected values.}
}
\value{
Modified \code{baseline} input.
}
\description{
Distort a baseline signal by offsetting a given range of values uniformly.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
