% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.sigmoid}
\alias{sim.sigmoid}
\title{Signal perturbation: sigmoid curve}
\usage{
sim.sigmoid(baseline, range, range_by_time = TRUE, shift,
  right_side = TRUE)
}
\arguments{
\item{baseline}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{range}{indices of affected values. Either a start and end index or the whole range. If length of \code{range} is 2 and second value is infinity, the second value is converted to highest possible value. If \code{range} is \code{NULL}, maximum range (1 to max) is taken.}

\item{range_by_time}{Boolean, indicates whether values in \code{range} correspond to value of \code{time} of signal, rather than row indices. This is recommended, since time values for fluorescence signals with non-uniform flow rate are also not uniform.}

\item{shift}{offset of the newly elevated/decreased level of the baseline.}

\item{right_side}{Boolean, increases whether the shift should occur over increasing time (versus decreasing).}
}
\value{
Modified \code{baseline} input.
}
\description{
Distort a baseline signal by permanently shifting part of the baseline by a specified amount, inserting a sigmoid curve as transition to the new part of the baseline.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
