% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.smooth_flex}
\alias{sim.smooth_flex}
\title{Signal perturbation: smooth flex}
\usage{
sim.smooth_flex(baseline, range, range_by_time = TRUE, strength,
  spikiness, D = c("gauss"))
}
\arguments{
\item{baseline}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{range}{indices of affected values. Either a start and end index or the whole range. If length of \code{range} is 2 and second value is infinity, the second value is converted to highest possible value. If \code{range} is \code{NULL}, maximum range (1 to max) is taken.}

\item{range_by_time}{Boolean, indicates whether values in \code{range} correspond to value of \code{time} of signal, rather than row indices. This is recommended, since time values for fluorescence signals with non-uniform flow rate are also not uniform.}

\item{strength}{maximum factor by which to stretch the selected values.}

\item{spikiness}{measure of how needle-shaped (vs. bell-shaped) the perturbation should be: values between 0 and 1.}

\item{D}{distribution from which to sample. Can be \code{"gauss"} (Gauss).}
}
\value{
Modified \code{baseline} input.
}
\description{
Distort a baseline signal by stretching or contracting the signal values (independently of time), effectively changing their variance, by factors generated by a probability density function.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
