% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.smooth_push}
\alias{sim.smooth_push}
\title{Signal perturbation: smooth push}
\usage{
sim.smooth_push(baseline, range, range_by_time = TRUE, strength,
  spikiness, D = c("gauss", "chisq", "-chisq"), D_par = c(3, 3))
}
\arguments{
\item{baseline}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{range}{indices of affected values. Either a start and end index or the whole range. If length of \code{range} is 2 and second value is infinity, the second value is converted to highest possible value. If \code{range} is \code{NULL}, maximum range (1 to max) is taken.}

\item{range_by_time}{Boolean, indicates whether values in \code{range} correspond to value of \code{time} of signal, rather than row indices. This is recommended, since time values for fluorescence signals with non-uniform flow rate are also not uniform.}

\item{strength}{maximum value of cell offset.}

\item{spikiness}{measure of how needle-shaped (vs. bell-shaped) the perturbation should be: values between 0 and 1.}

\item{D}{distribution from which to sample. Can be \code{"gauss"} (Gauss), \code{"chisq"} (chi-squared), \code{"-chisq"} (\code{"chisq"} flipped along the time axis).}

\item{D_par}{parameter for probability density function of given distribution. For \code{"chisq"} and \code{"-chisq"}, this is a vector of (1) degrees of freedom and (2) non-centrality parameter. (See function \code{dchisq}.)}
}
\value{
Modified \code{baseline} input.
}
\description{
Distort a baseline signal by offsetting a given range of values by adding values generated by a probability density function.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
