% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbations.R
\name{sim.stretch}
\alias{sim.stretch}
\title{Signal manipulation: stretch}
\usage{
sim.stretch(baseline, n = 1, prob = NULL)
}
\arguments{
\item{baseline}{baseline, as generated by function \code{sim.baseline}, possibly perturbed by other functions.}

\item{n}{number of time points which should be inserted after each original time point. This can either be a single integer, or a vector of integers from which values are drawn randomly.}

\item{prob}{vector of probabilities. If \code{n} is a vector: if \code{prob} is \code{NULL}, probability distribution over \code{n} is uniform, otherwise if \code{prob} is a vector of values, these are the corresponding probabilities for drawing each element.}
}
\value{
Modified \code{baseline} input.
}
\description{
Stretches out signal data with respect to time. Signal values are interlaced such that 0 or more \code{NA}s are inserted after each signal value and additional time points are generated accordingly.
}
\seealso{
See functions beginning with \code{sim.sample.} for examples of usage.
}
